Included ccignore file reading
This commit is being made to add functionality to ignore files at a
repository specific level. `.ccignore` is a file which should contain
**filenames** or **file types** to be ignored by the commit-comments
scripts.

Files included in the ccignore file will not have their contents
searched for `@commit` keyword, and will not be included in a commit
message.

If no ccignore file is present then no files will be ignored