More explanations in README & ignore file fix
The `.ccignore` file was reading the contents of the README. So the full
filename (name + path) needed to be added. It might have been a result
of a bug in filename/type matching in the script.

README was simplified and expanded in important areas. Formatting,
wording, and new sections. Added special thanks to Bryan Wyat.