Fix GNU grep compatibility.
This commit resolves an issue with a GNU grep command using the -P flag.
As a result, users without GNU Grep could not use the Perl Regular
Expressions flag.

This command has been removed and replaced with pcregrep which yields
the same results and also allows forward look-aheads.

Additionally, the test-files have been added to the .ccignore file so
cloning the repository does not cause users to have the test-files in
their first commit message.

Closes Issue #1