Portability for pcregrep in post-commit hook
This commit is being made to address portability for users without
pcregrep as a default package on their system. Some linux distributions
do not ship with pcregrep.

Implementation is borrowed from `post-commit-msg` hook which switches
grep commands depending on availability of pcregrep.

Assumption being made is that if a user does not have pcregrep they will
have GNU grep. Scripts do not suppose any other grep-like commands. An
update will be made to the dependencies section of the project.